feat: add ToPagin by Id CoreService [NotTested], add coreservice interface, TODO: immigrate from prev arch and use CoreLayer.CoreService| TODO: feat: ToPaging by orderfield, filter by key and value in db
Please enter the commit message for your changes. Lines starting

+++ CoreLayer/Utils/GridData.cs
@@ -9,10 +9,10 @@ namespace CoreLayer.Utils
     public class GridData<T> where T : class
     {
         public List<T> Data { get; set; } = new List<T>();
-        public long pageNumber { get; set; } = 1;
-        public long pageSize { get; set; } = long.MaxValue;
-        public long totalPages { get; set; }
-        public long totalCount { get; set; }
+        public int pageNumber { get; set; } = 1;
+        public int pageSize { get; set; } = int.MaxValue;
+        public int totalPages { get; set; }
+        public int totalCount { get; set; }
         public List<Filter> filters { get; set; } = new List<Filter>();
     }
 }
